+++ src/Stormancer.Plugins/Limits/Stormancer.Server.Plugins.Limits/LimitsController.cs
@@ -7,6 +7,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reactive.Linq;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace Stormancer.Server.Plugins.Limits
@@ -14,7 +15,7 @@ namespace Stormancer.Server.Plugins.Limits
     /// <summary>
     /// Provides S2S API for limits
     /// </summary>
-    [Service(Named =false, ServiceType = "stormancer.plugins.limits")]
+    [Service(Named = false, ServiceType = "stormancer.plugins.limits")]
     public class LimitsController : ControllerBase
     {
         private readonly ILimits limits;
@@ -60,6 +61,7 @@ namespace Stormancer.Server.Plugins.Limits
         private readonly IServiceLocator locator;
         private readonly ISerializer serializer;
         private readonly LimitsClientCache cache;
+        private readonly LimitsProxy limitsProxy;
 
         /// <summary>
         /// Create a new instance of the class.
@@ -68,34 +70,28 @@ namespace Stormancer.Server.Plugins.Limits
         /// <param name="locator"></param>
         /// <param name="serializer"></param>
         /// <param name="cache"></param>
-        public LimitsClient(RpcService rpc, IServiceLocator locator, ISerializer serializer, LimitsClientCache cache)
+        /// <param name="limitsProxy"></param>
+        public LimitsClient(RpcService rpc, IServiceLocator locator, ISerializer serializer, LimitsClientCache cache, LimitsProxy limitsProxy)
         {
             this.rpc = rpc;
             this.locator = locator;
             this.serializer = serializer;
             this.cache = cache;
+            this.limitsProxy = limitsProxy;
         }
 
         /// <summary>
         /// Gets connection limits state.
         /// </summary>
         /// <returns></returns>
-        public Task<UserConnectionLimitStatus> GetConnectionLimitStatus()
+        public Task<UserConnectionLimitStatus> GetConnectionLimitStatus(CancellationToken cancellationToken)
         {
-            return cache.Value.Get("limits", _ =>
-            {
-                return Retries.Retry<UserConnectionLimitStatus>(async (_) =>
-               {
-                   var sceneId = await locator.GetSceneId("stormancer.authenticator", string.Empty);
-                   return await rpc.Rpc("Limits.GetConnectionLimitStatus", new MatchSceneFilter(sceneId), s => { }, PacketPriority.MEDIUM_PRIORITY).Select(p =>
-                   {
-                       using (p)
-                       {
-                           return serializer.Deserialize<UserConnectionLimitStatus>(p.Stream);
-                       }
-                   }).SingleOrDefaultAsync();
-               }, RetryPolicies.ConstantDelay(4, TimeSpan.FromMilliseconds(500)), default, ex => true);
-            }, TimeSpan.FromSeconds(1));
+            return cache.Value.Get("limits", _ =>  Retries.Retry<UserConnectionLimitStatus>(
+                    (_) => limitsProxy.GetConnectionLimitStatus(cancellationToken), 
+                    RetryPolicies.ConstantDelay(4, TimeSpan.FromMilliseconds(500)), 
+                    cancellationToken, 
+                    ex => true)
+            , TimeSpan.FromSeconds(1));
 
         }
     }